requirements.txt
fastapi
uvicorn
pydantic
python-multipart
typing
google-genai